chore: Intialization of the 'Hello World' project for Keil V5
This is a initialization of the 'Hello World' project for the TM4C123 microcontroller. The printf function sends messages through the UART
peripheral. The serial output can be seen in the PC with the program Tera Term. This program is usefull to test small code snippets and watch the assembler instruction in the listing file.